Continue work on gutting Consul, Etcd
Apps are unable to deploy because of blobstore issue, I believe. BBS
throws an error stating that "failed to download cached artifacts".

Could still be a Consul issue, where some process is still expecting a
Consul lock but I haven't found it yet. Nothing is 404'ing on the
blobstore and I don't see any issues that immediately stand out.