Don't bail when moving non-existant certs
Fixes environment check when not using service discovery.  The check
hook tries to move certs to a new location (bug introduced in 1.8.0, but
bails if it doesn't exist, rather than just ignores it.

This `ignore` behaviour is required for certs that only exist when a
feature is enabled, such as the `dns-service-discovery` feature.

If the original behaviour is required, call relocate_secret as such:

```
relocate_secret <src> <dst> || env_ok=no
```