Added dynamically created instance counts manifest
Since we don't want to have a file with the default instance count
values defined (as we inherit them from whatever upstream cf-deployments
specifies), we only want to set them if the env file defines them.
Furthermore, we need to map the v1.x instance groups to the instance
groups defined by cf-deployments.

To this end, we create a dummy placeholder manifest fragment that gets
dynamically written during the blueprint hook, persisting for the
creation of the manifest, that contains only the go-patch replace
operations for specified `params.*_instances` values.