Defer the Cloud Config validation
Because we don't know what upstream extensions, networks or vm types are
going to be used, we now defer the cloud config checks to after the
manifest is generated and check the values referenced in the manifest
with those available in the cloud config in the pre-deploy hook.

Also improves output format and uses stderr in check and pre-deploy
hooks, and requires Genesis v2.7.23.