Suppress error when detecting external_db_user presence
If external_db_user is present, we need to warn users that they need to
set params.external_db_user to that value, as it is not picked up by
default in cf kit v2.x.  However, while it can be normal for that value
not to be present, the detection would log an extraneous warning that it
couldn't be found.  This fixes that issue.