Use genesis.env, not params.env
Once we updated the min genesi version to above 2.6.13, the
autopopulation of `params.env` has been removed.  Therefore, v1.10.2
causes genesis to complain that params.env exists, but won't work
without it.

The resolution is to use genesis.env instead of the outdated params.env
in the kit's manifest fragments, which is only needed in the blobstore
directory key settings.