Retain nats username when migrating v1 envs
This prevents availability outage when migrating from v1.10.x to 2.0.x.
For environments that have already migrated to v2, please use the
`v2-nats-credentials` feature to continue using the new credentials.
Failure to do so may result in a further availability outage during
deployments as the nats user is once again changes.