Refactor cloud-config: aws does not use statics
The original OCFP, which was targeting AWS, did not use static IPS for
the router and tcp-router (or anything else).  The previous change that
added statics for the router and tcp-router was not correct for AWS.
This change removes the statics for the router and tcp-router for AWS,
while keeping them for other IaaS providers.

If in the future we need to support static IPs for AWS, we can add a
feature such as `aws-static-edge-ips` to enable it, but for now this
change is sufficient to restore the original behavior.