Bump to v3.0.0 and require Genesis 3.1.0 stable
Well folks, we've finally made it to the big 3.0! This release
bumps the kit version to 3.0.0 and updates the minimum Genesis
requirement from 3.1.0-rc.9 to the stable 3.1.0 release.

Breaking changes:
- Removed support for azure, gcp, and vsphere platforms -- these are very much not updated for Genesis 3.1.0/OCFP yet
- Only aws, openstack, and stackit are supported for now

If you're on one of the removed platforms, you'll need to stick
with the 2.x series or migrate to a supported platform, or add the
functionalityy in yourself (it is open source, after all!).

The Genesis version bump means you'll need Genesis 3.1.0 or
later installed before you can use this kit. Time to upgrade
if you haven't already!